fix: 온보딩에서 멤버 정보를 불러와 404 에러뜨는 문제 해결 #249
