+++ frontend/src/components/SearchAddress/SearchAddress.styled.ts
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
 
-import IconButton from '../Button/IconButton';
-
 export const SearchAddressLayout = styled.div`
   background-color: ${({ theme }) => theme.colors.green50};
   position: fixed;
@@ -21,16 +19,4 @@ export const TopContainer = styled.div`
   width: 100%;
   gap: 2rem;
   padding: 2rem 3.6rem 4rem 3.6rem;
-`;
-
-export const ShortcutButtonContainer = styled.div`
-  display: flex;
-  gap: 2.4rem;
-  align-self: flex-start;
-`;
-
-export const IconButtonWrapper = styled(IconButton)`
-  display: flex;
-  gap: 0.4rem;
-  color: ${({ theme }) => theme.colors.green500};
 `;
