+++ frontend/src/components/SearchAddress/SearchAddress.types.ts
@@ -1,0 +1,10 @@
+export interface AddressMainProps {
+  title: string;
+  coordinateX: number;
+  coordinateY: number;
+  address?: string;
+}
+
+export interface AddressOnboardingProps extends AddressMainProps {
+  address: string;
+}
