+++ frontend/src/pages/MainPage/index.tsx
@@ -90,9 +90,7 @@ const MainPage = () => {
   };
 
   if (isSearchAddressOpen) {
-    return (
-      <SearchAddress onClose={toggleAddressSearch} onSelectAddressMain={updateAddress} isMain />
-    );
+    return <SearchAddress onClose={toggleAddressSearch} onSelectAddressMain={updateAddress} />;
   }
 
   return (
